Split notable player commands
